+++ render/deps.cljs
@@ -1,0 +1,6 @@
+{:npm-deps
+ {"@nextjournal/lang-clojure" "1.0.0"
+  "framer-motion" "7.6.19"
+  "react" "18.2.0"
+  "react-dom" "18.2.0"
+  "use-sync-external-store" "1.2.0"}}