+++ CHANGELOG.md
@@ -7,6 +7,20 @@ Changes can be:
 
 ## Unreleased
 
+...
+
+## 0.12.707 (2022-12-06)
+
+* 🐜 Make edn transmission not fail on bad keywords and symbols, fixes
+  #116
+* 🐜 Fix silent failure when analyzing invalid def, fixes #307
+* 🐞 Fixes an issue with codemirror syntax highlighting which
+  prevented multi-line strings to be displayed correctly
+* 🐞 Preserve whitespace in string viewer
+* 🐞 Fix parens placement when expanded string viewer is embedded in coll (#320)
+
+## 0.12.699 (2022-12-02)
+
 * 🌟 Clerk sync for vars holding atoms (#253, #268)
 
     Introduce `^:nextjournal.clerk/sync` metadata annotation for vars
@@ -29,7 +43,7 @@ Changes can be:
      `nextjournal.clerk.sci-env`.
     * Support alias resolution for `:render-fn`s (#276)
     * Upgrade to React 18.2 and introduce
-      `nextjournal.clerk.render.hoooks` as a thin cljs wrapper around
+      `nextjournal.clerk.render.hooks` as a thin cljs wrapper around
       [React hooks](https://reactjs.org/docs/hooks-intro.html) also
       useable from the sci env. (#237, #242)
     * Introduce `nextjournal.clerk.render.code` ns with support for
@@ -79,6 +93,8 @@ Changes can be:
     `:var-from-def?` key.
 
 * 💫 Make `nextjournal.clerk.parser` usable in CLJS
+
+* 💫 Support clearing the cache of a single result using `clerk/clear-cache!`
 
 * 💫 Set #-fragment when clicking on TOC items (works in unbundled
   case)